requirements.txt
streamlit
torch
torchvision
torchxrayvision
pillow
numpy
pydicom
opencv-python
grad-cam
scikit-image
matplotlib